pyproject.toml
[tool.poetry]
name = "drf_ghostpost_backend"
version = "0.1.0"
description = ""
authors = ["peyton glover <peyton.vx@protonmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
django = "^3.1.1"
djangorestframework = "^3.12.1"
django-cors-headers = "^3.5.0"
django-rest-multiple-models = "^2.1.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
